Split S7739 false-positives/index.ts into two exception modules
Separates the Yup/Joi validation-library exception (exception-libraries.ts)
from the structural/contract-based thenable exceptions (intentional-thenable.ts),
matching the multi-file false-positives convention used elsewhere (S6819,
S1848, S6767). index.ts is now a thin barrel re-exporting both.
